requirements.txt
numpy>=1.21
streamlit>=1.15
matplotlib>=3.5
shap>=0.41
coremltools>=6.0
pymoo>=0.5.0
torch>=1.13
scikit-learn>=1.0
pandas>=1.3
pyyaml>=5.4